requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.12.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
PySocks==1.7.1
python-telegram-bot==20.3
requests==2.31.0
schedule==1.2.1
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.4